DDF-3676 Refactor the registry IdentificationPlugin to query the catalog every time instead of caching registry ids
